Move wallet components into components folder
