+++ src/app/wallet/all/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { TransactionsBrowser } from './transactions-browser';
+import { TransactionsBrowser } from '../components/transactions-browser';
 
 export default function WalletAllPage() {
   return <TransactionsBrowser />;
