+++ src/app/wallet/components/filtered-transaction-table.tsx
